+++ Changelog.md
@@ -6,6 +6,7 @@
 - Add `/ub opt toggle <setting>` command to toggle boolean values
 - Add support for DHS/VHS hotkeys for common settings
 - Fix some memory leaks
+- Add `VTank.FixPortalLoops` setting that will attempt to detect and fix infinite portal loops while naving
 
 ## 0.1.2 (2020-01-02) [UtilityBeltInstaller-0.1.2.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/b1bcdf8401543ba32afd4cf480ff6408/UtilityBeltInstaller-0.1.2.exe)
 - "Fix" `/ub playeroption` on ACE, by sending the entire PlayerSystem packet